Split core/admin.py into multiple files that align with corresponding models files.
